Switch to using keys without hyphens in them
Switched to using keys without hyphens in them

When the increment function gets passed the key
`item-one`, the key is passed without quotes which
makes it appears as a non-string value to JavaScript.

It ends up reading as increment(item - one), which are
two separate variables, that are both undefined.

This throws the Uncaught ReferenceError item is not defined
error in the console whenever we try to add items to the cart.

For the original to have worked, we should have called the function
in the HTML thus: increment('${id}') to make sure it is being
rendered as a string